+++ lib/udap_security_test_kit/client_suite/token_request_verification.rb
@@ -1,5 +1,5 @@
 module UDAPSecurityTestKit
-  module TokenVerification
+  module TokenRequestVerification
     def verify_token_requests(oauth_flow)
       registration_token =
         begin
