+++ spec/udap_security_test_kit/client_suite/mock_udap_server_spec.rb
@@ -1,7 +1,7 @@
 RSpec.describe UDAPSecurityTestKit::MockUDAPServer, :request, :runnable do # rubocop:disable RSpec/SpecFilePathFormat
   include UDAPSecurityTestKit::URLs
   let(:suite_id) { 'udap_security_client' }
-  let(:test) { suite.children[1].children[0] } # access test
+  let(:test) { suite.children[2].children[0] } # access test
   let(:results_repo) { Inferno::Repositories::Results.new }
   let(:dummy_result) { repo_create(:result, test_session_id: test_session.id) }
   let(:reg_url) { client_registration_url }
